server/requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
aniso8601==9.0.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
bidict==0.22.1
cachelib==0.10.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
colorama==0.4.6
cryptography==38.0.3
distlib==0.3.6
dlib==19.24.0
fastapi==0.95.1
filelock==3.8.0
Flask==2.1.2
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask-SocketIO==5.3.3
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
httptools==0.5.0
idna==3.4
importlib-metadata==6.0.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
lxml==4.8.0
MarkupSafe==2.1.2
multidict==6.0.4
mysql-connector-python-rf==2.2.2
numpy==1.24.2
openai==0.27.6
opencv-python==4.6.0.66
pandas==1.4.2
Pillow==9.2.0
platformdirs==2.5.4
psycopg2==2.9.6
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.7
PyMySQL==1.0.2
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.4.0
python-socketio==5.8.0
pytube==15.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.11
starlette==0.26.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.16.7
watchfiles==0.19.0
webencodings==0.5.1
websockets==11.0.1
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
